Migrate typography components to MDX
