+++ src/typography/bodyStrong/BodyStrong.story.mdx
@@ -1,0 +1,33 @@
+import { Meta, Preview, Props, Story } from '@storybook/addon-docs/blocks'
+
+import { color } from '../../_utils/branding'
+import { TextBodyStrong } from '.'
+
+<Meta title="Design Tokens|Typography/" />
+
+# **TextBodyStrong**
+
+<Preview>
+  <Story name="TextBodyStrong">
+    <React.Fragment>
+      <h1>
+        <TextBodyStrong isInverted={false}>
+          {'The quick\n brown fox jumps\n over the lazy\n dog'}
+        </TextBodyStrong>
+      </h1>
+      <h1 style={{ backgroundColor: color.blue }}>
+        <TextBodyStrong isInverted>The quick brown fox jumps over the lazy dog</TextBodyStrong>
+      </h1>
+    </React.Fragment>
+  </Story>
+</Preview>
+
+## Usage
+
+```jsx
+import { TextBodyStrong } from '@blablacar/ui-library/build/typography/bodyStrong'
+
+<TextBodyStrong>The quick brown fox jumps over the lazy dog</TextBodyStrong>
+```
+
+<Props of={TextBodyStrong} />
