+++ src/typography/bodyStrong/BodyStrong.style.tsx
@@ -1,0 +1,8 @@
+import styled from 'styled-components'
+
+import { fontWeight } from '../../_utils/branding'
+import { TextBody } from '../body'
+
+export const TextBodyStrong = styled(TextBody)`
+  font-weight: ${fontWeight.medium};
+`
