+++ src/typography/bodyStrong/index.tsx
@@ -1,8 +1,1 @@
-import styled from 'styled-components'
-
-import { fontWeight } from '../../_utils/branding'
-import { TextBody } from '../body'
-
-export const TextBodyStrong = styled(TextBody)`
-  font-weight: ${fontWeight.medium};
-`
+export { TextBodyStrong } from './BodyStrong.style'
