--- src/typography/bodyStrong/story.tsx
@@ -1,24 +1,0 @@
-import React from 'react'
-import { boolean, text } from '@storybook/addon-knobs'
-import { storiesOf } from '@storybook/react'
-
-import { color } from '../../_utils/branding'
-import { BaseSection as Section } from '../../layout/section/baseSection'
-import { TextBodyStrong } from './index'
-
-const stories = storiesOf('Design Tokens|Typography/', module)
-
-stories.add('TextBodyStrong', () => (
-  <Section>
-    <h1>
-      <TextBodyStrong isInverted={boolean('isInverted', false)}>
-        {text('Text1', 'The quick\n brown fox jumps\n over the lazy\n dog')}
-      </TextBodyStrong>
-    </h1>
-    <h1 style={{ backgroundColor: color.blue }}>
-      <TextBodyStrong isInverted>
-        {text('Text2', 'The quick brown fox jumps over the lazy dog')}
-      </TextBodyStrong>
-    </h1>
-  </Section>
-))
