+++ src/typography/display2/Display2.story.mdx
@@ -1,0 +1,33 @@
+import { Meta, Preview, Props, Story } from '@storybook/addon-docs/blocks'
+
+import { color } from '../../_utils/branding'
+import { TextDisplay2 } from '.'
+
+<Meta title="Design Tokens|Typography/" />
+
+# **TextDisplay2**
+
+<Preview>
+  <Story name="TextDisplay2">
+    <React.Fragment>
+      <h1>
+        <TextDisplay2 isInverted={false}>
+          {'The quick\n brown fox jumps\n over the lazy\n dog'}
+        </TextDisplay2>
+      </h1>
+      <h1 style={{ backgroundColor: color.blue }}>
+        <TextDisplay2 isInverted>The quick brown fox jumps over the lazy dog</TextDisplay2>
+      </h1>
+    </React.Fragment>
+  </Story>
+</Preview>
+
+## Usage
+
+```jsx
+import { TextDisplay2 } from '@blablacar/ui-library/build/typography/display2'
+
+<TextDisplay2>The quick brown fox jumps over the lazy dog</TextDisplay2>
+```
+
+<Props of={TextDisplay2} />
