+++ src/typography/subHeaderStrong/SubHeaderStrong.story.mdx
@@ -1,0 +1,33 @@
+import { Meta, Preview, Props, Story } from '@storybook/addon-docs/blocks'
+
+import { color } from '../../_utils/branding'
+import { TextSubHeaderStrong } from '.'
+
+<Meta title="Design Tokens|Typography/" />
+
+# **TextSubHeaderStrong**
+
+<Preview>
+  <Story name="TextSubHeaderStrong">
+    <React.Fragment>
+      <h1>
+        <TextSubHeaderStrong isInverted={false}>
+          {'The quick\n brown fox jumps\n over the lazy\n dog'}
+        </TextSubHeaderStrong>
+      </h1>
+      <h1 style={{ backgroundColor: color.blue }}>
+        <TextSubHeaderStrong isInverted>The quick brown fox jumps over the lazy dog</TextSubHeaderStrong>
+      </h1>
+    </React.Fragment>
+  </Story>
+</Preview>
+
+## Usage
+
+```jsx
+import { TextSubHeaderStrong } from '@blablacar/ui-library/build/typography/subHeaderStrong'
+
+<TextSubHeaderStrong>The quick brown fox jumps over the lazy dog</TextSubHeaderStrong>
+```
+
+<Props of={TextSubHeaderStrong} />
