+++ src/typography/subHeaderStrong/SubHeaderStrong.style.tsx
@@ -1,0 +1,8 @@
+import styled from 'styled-components'
+
+import { fontWeight } from '../../_utils/branding'
+import { TextSubHeader } from '../subHeader'
+
+export const TextSubHeaderStrong = styled(TextSubHeader)`
+  font-weight: ${fontWeight.medium};
+`
