+++ src/typography/subHeaderStrong/index.tsx
@@ -1,8 +1,1 @@
-import styled from 'styled-components'
-
-import { fontWeight } from '../../_utils/branding'
-import { TextSubHeader } from '../subHeader'
-
-export const TextSubHeaderStrong = styled(TextSubHeader)`
-  font-weight: ${fontWeight.medium};
-`
+export { TextSubHeaderStrong } from './SubHeaderStrong.style'
