+++ src/typography/title/Title.style.tsx
@@ -1,0 +1,11 @@
+import styled from 'styled-components'
+
+import { color, font, fontWeight } from '../../_utils/branding'
+import { Text } from '../index'
+
+export const TextTitle = styled(Text)`
+  color: ${props => (props.isDisabled ? color.gray : color.midnightGreen)};
+  font-size: ${font.m.size};
+  font-weight: ${fontWeight.regular};
+  line-height: ${font.m.lineHeight};
+`
