+++ src/typography/titleStrong/TitleStrong.story.mdx
@@ -1,0 +1,33 @@
+import { Meta, Preview, Props, Story } from '@storybook/addon-docs/blocks'
+
+import { color } from '../../_utils/branding'
+import { TextTitleStrong } from '.'
+
+<Meta title="Design Tokens|Typography/" />
+
+# **TextTitleStrong**
+
+<Preview>
+  <Story name="TextTitleStrong">
+    <React.Fragment>
+      <h1>
+        <TextTitleStrong isInverted={false}>
+          {'The quick\n brown fox jumps\n over the lazy\n dog'}
+        </TextTitleStrong>
+      </h1>
+      <h1 style={{ backgroundColor: color.blue }}>
+        <TextTitleStrong isInverted>The quick brown fox jumps over the lazy dog</TextTitleStrong>
+      </h1>
+    </React.Fragment>
+  </Story>
+</Preview>
+
+## Usage
+
+```jsx
+import { TextTitleStrong } from '@blablacar/ui-library/build/typography/titleStrong'
+
+<TextTitleStrong>The quick brown fox jumps over the lazy dog</TextTitleStrong>
+```
+
+<Props of={TextTitleStrong} />
