+++ src/typography/titleStrong/TitleStrong.style.tsx
@@ -1,0 +1,8 @@
+import styled from 'styled-components'
+
+import { fontWeight } from '../../_utils/branding'
+import { TextTitle } from '../title'
+
+export const TextTitleStrong = styled(TextTitle)`
+  font-weight: ${fontWeight.medium};
+`
