+++ src/typography/titleStrong/index.tsx
@@ -1,8 +1,1 @@
-import styled from 'styled-components'
-
-import { fontWeight } from '../../_utils/branding'
-import { TextTitle } from '../title'
-
-export const TextTitleStrong = styled(TextTitle)`
-  font-weight: ${fontWeight.medium};
-`
+export { TextTitleStrong } from './TitleStrong.style'
