+++ src/typography/typography.story.mdx
@@ -98,42 +98,42 @@ _Text Components_ are text based components.
 `TextComponents` are exported individually.
 
 ```js
-import TextBody from './typography/body'
+import { TextBody } from './typography/body'
 <TextBody>The quick brown fox jumps over the lazy dog</TextBody>
 ```
 
 ```js
-import TextBodyStrong from './typography/bodyStrong'
+import { TextBodyStrong } from './typography/bodyStrong'
 <TextBodyStrong>The quick brown fox jumps over the lazy dog</TextBodyStrong>
 ```
 
 ```js
-import TextCaption from './typography/caption'
+import { TextCaption } from './typography/caption'
 <TextCaption>The quick brown fox jumps over the lazy dog</TextCaption>
 ```
 
 ```js
-import TextTitle from './typography/title'
+import { TextTitle } from './typography/title'
 <TextTitle>The quick brown fox jumps over the lazy dog</TextTitle>
 ```
 ```js
-import TextTitleStrong from './typography/titleStrong'
+import { TextTitleStrong } from './typography/titleStrong'
 <TextTitleStrong>The quick brown fox jumps over the lazy dog</TextTitleStrong>
 ```
 ```js
-import TextSubHeader from './typography/subHeader'
+import { TextSubHeader } from './typography/subHeader'
 <TextSubHeader>The quick brown fox jumps over the lazy dog</TextSubHeader>
 ```
 ```js
-import TextSubHeaderStrong from './typography/subHeaderStrong'
+import { TextSubHeaderStrong } from './typography/subHeaderStrong'
 <TextSubHeaderStrong>The quick brown fox jumps over the lazy dog</TextSubHeaderStrong>
 ```
 ```js
-import TextDisplay2 from './typography/display2'
+import { TextDisplay2 } from './typography/display2'
 <TextDisplay2>The quick brown fox jumps over the lazy dog</TextDisplay2>
 ```
 ```js
-import TextDisplay1 from './typography/display1'
+import { TextDisplay1 } from './typography/display1'
 <TextDisplay1>The quick brown fox jumps over the lazy dog</TextDisplay1>
 ```
